requirements.txt
aalpy
appdirs
attrs
colorama
crcmod
filelock
fsspec
future
Jinja2
lxml
MarkupSafe
mpmath
nest-asyncio
networkx
numpy
packaging
pillow
posix-ipc
psutil
pydot
pyparsing
pyshark
scapy
scikit-learn
sympy
sysv-ipc
termcolor
toml
torch
torchaudio
torchvision
triton
typing-extensions
